+++ Documentation/Release Notes.md
@@ -11,7 +11,12 @@
 ## STH Test
 
 - The EEPROM tests now also writes the firmware release name (`STH` → `Firmware` → `Release Name` in the [configuration](../Configuration/config.yaml)) into the EEPROM.
-- The EEPROM test now resets the under voltage counter to `0`.
+- The EEPROM test now sets the
+
+  - under voltage counter, and the
+  - watchdog reset counter
+
+  to `0`.
 
 # Version 1.0.4